requirements.txt
flask==3.0.2
flask-bcrypt==1.0.1
scikit-learn==1.4.1.post1
spacy==3.7.4
pdfplumber==0.11.0
python-docx==1.1.0
pandas==2.2.1
joblib==1.3.2
numpy==1.26.4